backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.3
numpy==1.24.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Machine Learning (temporarily disabled for lighter deployment)
# scikit-learn==1.3.2
# xgboost==2.0.2
# joblib==1.3.2

# Feature Engineering
ta==0.10.2
# talib==0.4.0  # Temporarily disabled - complex compilation

# Cache & Queue
redis==5.0.1
celery==5.3.4

# Monitoring
prometheus-client==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1